refactor: moving the API Management clients
